state sync: address comments on messages
